+++ tests/ProviderLoginCheckWrapperTest.php
@@ -6,7 +6,6 @@ use seregazhuk\PinterestBot\Api\Request;
 use seregazhuk\PinterestBot\Api\Response;
 use seregazhuk\PinterestBot\Api\CurlAdapter;
 use seregazhuk\PinterestBot\Api\Providers\Provider;
-use seregazhuk\PinterestBot\Exceptions\AuthException;
 use seregazhuk\PinterestBot\Exceptions\InvalidRequestException;
 use seregazhuk\PinterestBot\Api\Providers\ProviderLoginCheckWrapper;
 
@@ -16,11 +15,10 @@ class ProviderLoginCheckWrapperTest extends \PHPUnit_Framework_TestCase
      * For not logged in request.
      *
      * @test
+     * @expectedException seregazhuk\PinterestBot\Exceptions\AuthException
      */
     public function failWhenLoginIsRequired()
     {
-        $this->expectException(AuthException::class);
-
         $provider = new TestProvider(new Request(new CurlAdapter()), new Response());
         $wrapper = new ProviderLoginCheckWrapper($provider);
         $wrapper->testFail();
@@ -34,10 +32,12 @@ class ProviderLoginCheckWrapperTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('success', $wrapper->testSuccess());
     }
 
-    /** @test */
+    /**
+     * @test
+     * @expectedException seregazhuk\PinterestBot\Exceptions\InvalidRequestException
+     */
     public function callNonexistentMethod()
     {
-        $this->expectException(InvalidRequestException::class);
         $provider = new TestProvider(new Request(new CurlAdapter()), new Response());
         $wrapper = new ProviderLoginCheckWrapper($provider);
         $wrapper->badMethod();